# internal/features/tasks/transport/http/patch_task.go
package tasks_transport_http

import (
	"fmt"
	"net/http"
	"time"

	"github.com/med0viy/practika/internal/core/domain"
	core_logger "github.com/med0viy/practika/internal/core/logger"
	core_http_request "github.com/med0viy/practika/internal/core/transport/http/request"
	core_http_response "github.com/med0viy/practika/internal/core/transport/http/response"
	core_http_types "github.com/med0viy/practika/internal/core/transport/http/types"
)

type PatchTaskRequest struct {
	Title       core_http_types.Nullable[string]               `json:"title"         swaggertype:"string"   example:"Погулять c собакой"`
	Description core_http_types.Nullable[string]               `json:"description"   swaggertype:"string"   example:"Одеть собаке шлейку и пойти c ней в парк"`
	Complited   core_http_types.Nullable[bool]                 `json:"complited"     swaggertype:"boolean"  example:"false"`
	IsImportant core_http_types.Nullable[bool]                 `json:"is_important"  swaggertype:"boolean"  example:"true"`
	IsInMyDay   core_http_types.Nullable[bool]                 `json:"is_in_my_day"  swaggertype:"boolean"  example:"false"`
	DueDate     core_http_types.Nullable[core_http_types.Date] `json:"due_date"      swaggertype:"string"   example:"2026-04-03"`
	ListID      core_http_types.Nullable[int]                  `json:"list_id"       swaggertype:"integer"  example:"31"`
}

type PatchTaskResponse TaskDTOResponse

// PatchTask      godoc
// @Summary       Измененить задачу
// @Description   Измение информации об уже существующей в системе задачи
// @Description   ### Логика обновления полей (Three-state logic):
// @Description   1. **Поле не передано**: `description` игнорируется, значение в БД не меняется
// @Description   2. **Явно передано значение**: `"description": "Одеть собаке шлейку и пойти с ней в парк"` - устанавливает новое описание в БД
// @Description   3. **Передан null**: `"description": null` - очищает поле в БД (set to NULL)
// @Tags          tasks
// @Accept        json
// @Produce       json
// @Param         id path int true "ID изменяемой задачи"
// @Param         request body PatchTaskRequest true "PatchTask тело запроса"
// @Success       200 {object} PatchTaskResponse "Успешно изменнённая задача"
// @Failure       400 {object} core_http_response.ErrorResponse "Bad request"
// @Failure       404 {object} core_http_response.ErrorResponse "Task not found"
// @Failure       409 {object} core_http_response.ErrorResponse "Сonflict"
// @Failure       500 {object} core_http_response.ErrorResponse "Internal server error"
// @Router        /tasks/{id} [patch]
func (r *PatchTaskRequest) Validate() error {
	if r.Title.Set {
		if r.Title.Value == nil {
			return fmt.Errorf("`Title` can't be NULL")
		}

		titleLen := len([]rune(*r.Title.Value))
		if titleLen < 1 && titleLen > 100 {
			return fmt.Errorf("`Title` must be between 1 and 100 symbols")
		}
	}

	if r.Description.Set {
		if r.Description.Value != nil {
			descriptionLen := len([]rune(*r.Description.Value))
			if descriptionLen < 1 && descriptionLen > 1000 {
				return fmt.Errorf("`Description` must be between 1 and 1000 symbols")
			}
		}
	}

	if r.Complited.Set {
		if r.Complited.Value == nil {
			return fmt.Errorf("`Complited` can't be NULL")
		}
	}

	if r.IsImportant.Set {
		if r.IsImportant.Value == nil {
			return fmt.Errorf("`IsImportant` can't be NULL")
		}
	}

	if r.IsInMyDay.Set {
		if r.IsInMyDay.Value == nil {
			return fmt.Errorf("`IsInMyDay` can't be NULL")
		}
	}

	return nil
}

func (h *TasksHTTPHandler) PatchTask(w http.ResponseWriter, r *http.Request) {
	ctx := r.Context()
	log := core_logger.LoggerContext(ctx)
	responseHandler := core_http_response.NewHTTPResponseHandler(log, w)

	taskId, err := core_http_request.GetIntPathValue(r, "id")
	if err != nil {
		responseHandler.ErrorResponse(
			err,
			"failed to get taskID path value",
		)

		return
	}

	var req PatchTaskRequest
	if err := core_http_request.DecodeAndValidateRequest(r, &req); err != nil {
		responseHandler.ErrorResponse(
			err,
			"failed to decode and validate HTTP request",
		)

		return
	}

	taskPatch := taskPatchFromRequest(req)

	taskDomain, err := h.tasksService.PatchTask(ctx, taskId, taskPatch)
	if err != nil {
		responseHandler.ErrorResponse(
			err,
			"failed to patch task",
		)

		return
	}

	response := PatchTaskResponse(TaskDTOFromDomain(taskDomain))

	responseHandler.JSONResponse(response, http.StatusOK)
}

func taskPatchFromRequest(request PatchTaskRequest) domain.TaskPatch {
	var domainDueDate domain.Nullable[time.Time]
	domainDueDate.Set = request.DueDate.Set

	if request.DueDate.Value != nil {
		t := (time.Time)(*request.DueDate.Value)
		domainDueDate.Value = &t
	}

	return domain.NewTaskPatch(
		request.Title.ToDomain(),
		request.Description.ToDomain(),
		request.Complited.ToDomain(),
		request.IsImportant.ToDomain(),
		request.IsInMyDay.ToDomain(),
		domainDueDate,
		request.ListID.ToDomain(),
	)
}