# internal/features/users/transport/http/dto_common.go
package users_transport_http

import "github.com/med0viy/practika/internal/core/domain"

type UserDTOResponse struct {
	ID          int     `json:"id"           example:"14"`
	Version     int     `json:"version"      example:"88"`
	FullName    string  `json:"full_name"    example:"Burmaldak Burmaldakovich"`
	PhoneNumber *string `json:"phone_nuber"  example:"+79998886742"`
}

func userDTOFromDomain(user domain.User) UserDTOResponse {
	return UserDTOResponse{
		ID:          user.ID,
		Version:     user.Version,
		FullName:    user.FullName,
		PhoneNumber: user.PhoneNumber,
	}
}

func usersDTOFromDomains(users []domain.User) []UserDTOResponse {
	usersDTO := make([]UserDTOResponse, len(users))

	for i, user := range users {
		usersDTO[i] = userDTOFromDomain(user)
	}

	return usersDTO
}